+++ tests/Integration/Resources/DSC_SqlProtocolTcpIp.Integration.Tests.ps1
@@ -102,7 +102,7 @@ Describe "$($script:dscResourceName)_Integration" -Tag @('Integration_SQL2016', 
             $resourceCurrentState.IpAddress | Should-Be $ConfigurationData.AllNodes.IpAddress
         }
 
-        It 'Should return $true when Test-DscConfiguration is run' {
+        It 'Should return True when Test-DscConfiguration is run' {
             Test-DscConfiguration -Verbose -ErrorAction 'Stop' | Should-Be 'True'
         }
     }