+++ tests/Integration/Resources/DSC_SqlRS.Integration.Tests.ps1
@@ -167,7 +167,7 @@ Describe "$($script:dscResourceName)_Integration" -Tag @('Integration_SQL2016', 
             $resourceCurrentState.ReportServerReservedUrl | Should-ContainCollection 'http://+:80'
         }
 
-        It 'Should return $true when Test-DscConfiguration is run' {
+        It 'Should return True when Test-DscConfiguration is run' {
             Test-DscConfiguration -Verbose -ErrorAction 'Stop' | Should-Be 'True'
         }